+++ assets/api/schema.py
@@ -34,16 +34,18 @@ class Assets(Schema):
     item_id: int
     name: str
     quantity: int
+    location_id: int
     location: str
     price: Any
 
 
 class Requests(Schema):
     id: int
-    order: Any
     status: str
+    order: Any
     action: str
     created: datetime
     closed: Optional[datetime] = None
     approver: Optional[Any]
     requestor: Any
+    actions: Any